+++ src/Util/GlideUrl.php
@@ -4,11 +4,14 @@
 namespace AmpedWeb\GlideInABox\Util;
 
 
+use AmpedWeb\GlideInABox\Traits\Adjustments;
 use AmpedWeb\GlideInABox\Traits\Border;
 use AmpedWeb\GlideInABox\Traits\Crop;
+use AmpedWeb\GlideInABox\Traits\Effects;
 use AmpedWeb\GlideInABox\Traits\Encode;
 use AmpedWeb\GlideInABox\Traits\Flip;
 use AmpedWeb\GlideInABox\Traits\Orientation;
+use AmpedWeb\GlideInABox\Traits\PixelDensity;
 use AmpedWeb\GlideInABox\Traits\Size;
 use Illuminate\Support\Str;
 use League\Glide\Urls\UrlBuilder;
@@ -16,7 +19,7 @@ use League\Glide\Urls\UrlBuilderFactory;
 
 class GlideUrl
 {
-    use Orientation, Flip, Crop, Size, Border, Encode;
+    use Orientation, Flip, Crop, Size, PixelDensity, Adjustments, Effects, Border, Encode;
 
     /**
      * The filepath of our image being manipulated